Refactor duplicate indicator detection logic to scanner.py
- Moved `_extract_touchup_text` and `_get_text_for_comparison` from `app_gui.py` to `scanner.py` as standalone functions.
- Updated `detect_indicators` in `scanner.py` to use these local functions and removed the dependency on `app_instance`.
- Updated `app_gui.py` to import `detect_indicators` from `scanner.py` and removed the duplicated methods.
- This improves code health by centralizing scanning logic in the `scanner` module.
